delete Api module
