+++ packages/xdl/src/Simulator.ts
@@ -10,9 +10,9 @@ import semver from 'semver';
 
 import {
   Analytics,
-  Api,
   BundleIdentifier,
   delayAsync,
+  downloadAppAsync,
   learnMore,
   Logger,
   NotificationCode,
@@ -424,7 +424,7 @@ export async function _downloadSimulatorAppAsync(
 
   fs.mkdirpSync(dir);
   try {
-    await Api.downloadAsync(url, dir, { extract: true }, downloadProgressCallback);
+    await downloadAppAsync(url, dir, { extract: true }, downloadProgressCallback);
   } catch (e) {
     fs.removeSync(dir);
     throw e;
