+++ packages/xdl/src/utils/downloadApkAsync.ts
@@ -1,7 +1,7 @@
 import fs from 'fs-extra';
 import path from 'path';
 
-import { Api, UserSettings, Versions } from '../internal';
+import { downloadAppAsync, UserSettings, Versions } from '../internal';
 
 function _apkCacheDirectory() {
   const dotExpoHomeDirectory = UserSettings.dotExpoHomeDirectory();
@@ -26,6 +26,6 @@ export async function downloadApkAsync(
     return apkPath;
   }
 
-  await Api.downloadAsync(url, apkPath, undefined, downloadProgressCallback);
+  await downloadAppAsync(url, apkPath, undefined, downloadProgressCallback);
   return apkPath;
 }
